Fix 3069: update for feedback
